auto-t: Call default instead of reload in teardown
This solves an issue where test_ignore_candidate_list_quirk() uses
set_value() to set a vendor element telling the station that the
candidate list should be ignored, and that config was never cleared in
teardown. This resulted in subsequent tests ignoring the candidate list
and doing a normal scan. This caused them to connect to a BSS other than
the one in the candidate list, causing the tests to fail.

When calling default(), it first resets all configs to default, then
calls reload(). So we still get the reload() call as before, with the
added benefit of also clearing the configs. This makes the tests pass.