auto-t: Disable MAC randomization for a few tests
The scan subsets now contains both 2.4GHz and 5GHz channels, this seems
to cause an issue when MAC randomization is used together with the
simulated hardware.

The AP(s) in these tests operate on 2.4GHz. From the logs it's clear
that the station's probe request to the 2.4GHz AP does arrive, and the
AP responds, but the station never processes the response. If the AP is
changed to 5GHz then everything works fine. I believe this is because
2.4GHz is scanned first, then there's a MAC change, then 5GHz is
scanned. For some reason, this causes the simulated hardware to miss the
response from the first scan, possibly because the MAC address doesn't
match.